:recycle: replace app.ts to index.ts (calling App)
