+++ package.json
@@ -11,7 +11,7 @@
     "yarn": "1.x"
   },
   "scripts": {
-    "start": "node dist/app.js",
+    "start": "node dist/index.js",
     "build": "tsc --outDir dist/",
     "typecheck": "tsc --noEmit"
   },
